cleanups and store values to db

+++ server/migrations/2020-06-14-185111_create_intraday_prices/up.sql
@@ -3,7 +3,7 @@ CREATE TABLE IF NOT EXISTS intraday_prices
     id        INTEGER        NOT NULL PRIMARY KEY,
     stock_id  INTEGER        NOT NULL,
     price     DECIMAL(12, 2) NOT NULL,
-    volume    INTEGER        NOT NULL,
+    volume    BIGINT         NOT NULL,
     timestamp TIMESTAMP      NOT NULL,
     FOREIGN KEY (stock_id) REFERENCES stocks (id)
 );
